requirements.txt
Babel==2.6.0
blinker==1.4
certifi==2018.4.16
chardet==3.0.4
click==6.7
coverage==4.5.1
decorator==4.3.0
defusedxml==0.5.0
dlib==19.13.1
face-recognition==1.2.2
face-recognition-models==0.3.0
Flask==0.12
Flask-Babel==0.11.2
Flask-Login==0.4.1
Flask-Mail==0.9.1
Flask-OpenID==1.2.5
Flask-Script==2.0.5
Flask-SQLAlchemy==2.3.2
Flask-WhooshAlchemy==0.56
Flask-WTF==0.14.2
flipflop==1.0
gevent==1.3.3
greenlet==0.4.13
guess-language==0.2
idna==2.7
itsdangerous==0.24
Jinja2==2.10
MarkupSafe==1.0
numpy==1.14.3
opencv-python==3.4.1.15
pbr==4.0.3
Pillow==5.1.0
PyMySQL==0.8.1
python3-openid==3.1.0
pytz==2018.4
requests==2.19.1
six==1.11.0
SQLAlchemy==1.2.8
sqlalchemy-migrate==0.11.0
sqlparse==0.2.4
Tempita==0.5.2
urllib3==1.23
uWSGI==2.0.17
Werkzeug==0.14.1
Whoosh==2.7.4
WTForms==2.2
